+++ mobile/src/composables/index.js
@@ -1,2 +1,0 @@
-// 全局可复用的 Vue 组合式函数
-// 使用方式：import { useXXX } from '@c/composables'