+++ mobile/src/constants/index.js
@@ -1,2 +1,0 @@
-// 全局常量/枚举配置
-// 使用方式：import { XXX } from '@c/constants'