+++ mobile/src/utils/index.js
@@ -1,2 +1,0 @@
-// 全局通用工具函数
-// 使用方式：import { xxx } from '@c/utils'